Docker/RPI3_Receiver/requirements.txt
influxdb
pyserial